Switch to using Thread's ContextClassLoader
- This is primarily to allow org.python.ImportLib to use the newl
JoinClassLoader in TestDaemon to load the new tests/temp/ classes.
Otherwise, it uses the ImportLib class' classloader, which is the
previous test case's classloader, which has the wrong set of classes for
tests/temp/.
- There should be no impact under non-testing circumstances as nobody
should be messing with the running Thread's ContextClassLoader.
- Fixes test cases which only work correctly on the first run, and which
may not work correctly for subsequent runs if the underlying classes are
replaced.