Implement float.__repr__
This is a *very* simplified version of what CPython does in
PyOS_double_to_string: use native string formatting and then fix it
up.
There are edge cases where this does not give the same result
as CPython's float.__repr__, but it should never be *wrong*.

This removes the need to normalize floats in the test output,
and it fixes the float modulo tests.