Respect @expectedFailure when processing `not_implemented`.
During the processing of `not_implemented`, the value of
`__unittest_expecting_failure__` was enforced, instead of marking it
when not already set.

Also, it turns out that the marker was set on the method, which was
shared among different testcases. The marker should instead be set on
the instance of the `TestCase`, which makes sure it is only remembered
for that specific run.