fix maximum stack size for an edge case The bytecode generated for the test case added in this commit reaches the limit of the stack size, see the ASM analysis here:
https://gist.github.com/eliasdorneles/742ccff84a0474e6904a369293ab1dd9

In the max_stack value calculation, len(self.parameters) was 0,
so that max_stack ends up with value 8.

Looking at the ASM analysis linked above, one can see that it
reaches the limit, and even though the INVOKESTATIC function
would consume 1 element from the stack (the argument) before
adding its result to the stack, this seems to not be enough.