Massive code reformat with IntelliJ IDEA
- Some small manual intervention to clear the ‘boolean expression could
be simplified’ errors, a misordered ‘static private’ access modifier,
some misplaced semicolons, overenthusiastic removal of blank lines, and
missing newlines at end of file.
- Unfortunately it applies the continuation indent setting equally to
method invocations which have the ending ) at the same level, as to
those which don’t, so it looks a little weird on the @org.python.Method
annotations.
- Disable two more checkstyle rules
  - EmptyBlock: there are a few finally blocks that just contain a
  commented out debug statement.
  - HideUtilityClassConstructor: it thinks org.Python is a utility class
  which ought to have a different access modifier.