Fix dict conversion errors for strings
When sets and frozen sets were converted to dicts, if they contained
string elements those string elements would not be converted as they
would in python. Fix this.

As part of this commit, the mixed case unit test was removed, this is
due to randomisation present in sets in python 3.4 causing flakey tests.

Signed-off-by: Meghan Halton <meghalton@gmail.com>
Signed-off-by: Pamela McA'Nulty <pamela@mcanulty.org>