Patch rich comparisons to handle Types
- The implementation I did in #122 assumes that when calling
__getattribute_null() on an operand to dynamically obtain the method, it
always returns a Method, which can be duly invoked for a result.
- However, this assumption is faulty for org.python.types.Type operands
(`type(1)`), as its __getattribute_null() returns a Function which isn’t
already bound to an object instance.
- In any case it only makes sense to invoke the comparison method on an
object, not on a class/type, so just avoid this entirely by testing if
it’s a Method that’s already bound to an object instance.