Handle return statements inside non-if structures
An implicit reutrn is required in a Python function even when there is
an explicit return, if the statement is inside a nested structure
(loop, condition, or try-catch block).

This commit adds a property that calculates whether there is a nested
structure inside a block, and use it to determine whether the block
needs an extra implicit return statement when visiting the explicit one.

The same logic is added to the visit_Lambda as well.

Fixes #533.