long and double needs an extra variable in locals array
This is a partial attempt to fix a bug triggered when using the
primitive type long in an argument for a method.

The issue is that long and double are treated different than
values of other types, basicaly they need extra space. The current
code seems to assume variables have the same size.

I don't fully understand the implications of the changes I'm trying
to make, so I'm defintely out of my depth here.