Fix ability to use long b'' literals
The BIPUSH opcode treats values as 8-bit signed, and sign-extends when
storing.  Using BIPUSH to set the length of a byte array limits the
length to 127 bytes.

Use ICONST_val instead of BIPUSH to select the appropriate instruction
for setting the length of a byte array.  Also, only pass values in the
8-bit signed range to BIPUSH.